Use int for token counts

+++ src/palimpzest/dataclasses.py
@@ -161,10 +161,10 @@ class RecordOpStats:
     generated_fields: Optional[List[str]] = None
 
     # the total number of input tokens processed by this operator; None if this operation did not use an LLM
-    total_input_tokens: float = 0.0
+    total_input_tokens: int = 0
 
     # the total number of output tokens processed by this operator; None if this operation did not use an LLM
-    total_output_tokens: float = 0.0
+    total_output_tokens: int = 0
 
     # the total cost of processing the input tokens; None if this operation did not use an LLM
     total_input_cost: float = 0.0
@@ -322,7 +322,7 @@ class ExecutionStats:
     """
 
     # string for identifying this workload execution
-    execution_id: str  |None= None
+    execution_id: str | None = None
 
     # dictionary of PlanStats objects (one for each plan run during execution)
     plan_stats: Dict[str, PlanStats] = field(default_factory=dict)
